修改模板.支持lombok

+++ src/main/java/com/demon/mybatis/generator/StringUtils.java
@@ -93,7 +93,7 @@ public class StringUtils {
         switch (databaseType) {
             case Types.DECIMAL:
             case Types.REAL:
-                colType = "Double";
+                colType = "BigDecimal";
                 break;
             case Types.BIGINT:
                 colType = "Long";
